Split up keyboard.rs into smaller chunks
