+++ crates/bevy_input/src/keyboard/event.rs
@@ -1,0 +1,10 @@
+use crate::keyboard::KeyCode;
+use crate::ElementState;
+
+/// A key input event from a keyboard device
+#[derive(Debug, Clone)]
+pub struct KeyboardInput {
+    pub scan_code: u32,
+    pub key_code: Option<KeyCode>,
+    pub state: ElementState,
+}
