+++ crates/bevy_input/src/keyboard/mod.rs
@@ -1,0 +1,5 @@
+pub mod event;
+pub mod key_code;
+pub mod system;
+
+pub use crate::keyboard::{event::*, key_code::*, system::*};
